share functionality and clean up
